+++ src/components/spartan/SBreadcrumbs/SBreadcrumbs.vue
@@ -1,11 +1,7 @@
 <template>
     <nav class="flex" aria-label="Breadcrumb">
         <ol role="list" class="flex items-center space-x-3">
-            <slot></slot>
+            <slot />
         </ol>
     </nav>
 </template>
-
-<script>
-export default {};
-</script>